Move create vc
