+++ packages/site/src/pages/index.tsx
@@ -29,7 +29,6 @@ import {
   configureGoogleAccount,
   connectHederaAccount,
   connectSnap,
-  createVC,
   createVP,
   deleteAllVCs,
   getCurrentNetwork,
@@ -44,6 +43,7 @@ import {
   verifyVP,
 } from '../utils';
 import { validHederaChainID } from '../utils/hedera';
+import CreateVC from './cards/CreateVC';
 import GetAllVCs from './cards/GetAllVCs';
 import GetCurrentDIDMethod from './cards/GetCurrentDIDMethod';
 import GetDID from './cards/GetDID';
@@ -56,9 +56,6 @@ const Index = () => {
   const [hederaAccountConnected, setHederaAccountConnected] = useState(false);
 
   const [hederaAccountId, setHederaAccountId] = useState('0.0.15215');
-
-  const [createVCName, setCreateVCName] = useState('Kiran Pachhai');
-  const [createVCNickname, setCreateVCNickname] = useState('KP Woods');
 
   const { vcId, setVcId, vc, setVc, vcIdsToBeRemoved, setVcIdsToBeRemoved } =
     useContext(VcContext);
@@ -143,38 +140,6 @@ const Index = () => {
     try {
       setCurrentChainId(await getCurrentNetwork());
       await togglePopups();
-    } catch (e) {
-      console.error(e);
-      dispatch({ type: MetamaskActions.SetError, payload: e });
-    }
-  };
-
-  const handleCreateVCClick = async () => {
-    try {
-      setCurrentChainId(await getCurrentNetwork());
-      const vcKey = 'profile';
-      const vcValue = {
-        name: createVCName,
-        nickname: createVCNickname,
-      };
-      const options = {
-        // If you want to auto save the generated VCs to multiple stores, you can pass an array like so:
-        // store: ['snap', 'googleDrive'],
-        store: 'snap',
-        returnStore: true,
-      };
-      const credTypes = ['ProfileNamesCredential'];
-      const saved = await createVC(vcKey, vcValue, options, credTypes);
-      const savedJson = JSON.parse(JSON.stringify(saved));
-      if (savedJson.length > 0) {
-        const vcIdsToAdd: any = [];
-        savedJson.forEach((data: any) => {
-          vcIdsToAdd.push(data.id);
-        });
-        setVcId(vcIdsToAdd.toString());
-        setVcIdsToBeRemoved(vcIdsToAdd.toString());
-        console.log('created and saved VC: ', saved);
-      }
     } catch (e) {
       console.error(e);
       dispatch({ type: MetamaskActions.SetError, payload: e });
@@ -459,53 +424,11 @@ const Index = () => {
         />
 
         {/* =============================================================================== */}
-        {(validHederaChainID(currentChainId) && hederaAccountConnected) ||
-        (!validHederaChainID(currentChainId) && !hederaAccountConnected) ? (
-          <Card
-            content={{
-              title: 'createVC',
-              description: 'Create and Save VerifiableCredential',
-              form: (
-                <form>
-                  <label>
-                    Enter your name
-                    <input
-                      type="text"
-                      style={{ width: '100%' }}
-                      value={createVCName}
-                      onChange={(e) => setCreateVCName(e.target.value)}
-                    />
-                  </label>
-                  <br />
-                  <label>
-                    Enter your nickname
-                    <input
-                      type="text"
-                      style={{ width: '100%' }}
-                      value={createVCNickname}
-                      onChange={(e) => setCreateVCNickname(e.target.value)}
-                    />
-                  </label>
-                </form>
-              ),
-              button: (
-                <SendHelloButton
-                  buttonText="Generate VC"
-                  onClick={handleCreateVCClick}
-                  disabled={!state.installedSnap}
-                />
-              ),
-            }}
-            disabled={!state.installedSnap}
-            fullWidth={
-              state.isFlask &&
-              Boolean(state.installedSnap) &&
-              !shouldDisplayReconnectButton(state.installedSnap)
-            }
-          />
-        ) : (
-          ''
-        )}
+        <CreateVC
+          currentChainId={currentChainId}
+          setCurrentChainId={setCurrentChainId}
+          hederaAccountConnected={hederaAccountConnected}
+        />
         {/* =============================================================================== */}
         {(validHederaChainID(currentChainId) && hederaAccountConnected) ||
         (!validHederaChainID(currentChainId) && !hederaAccountConnected) ? (
